Refactor explanation tests into their own package, one test for language element explanations (negations and disjunctions) and one for reasoner explanations where rules have been applied
